fix(ci): 修复 Kotlin 编译错误
3 个错误：

1. ChatScreen.kt:384 Unresolved reference 'animateFloat'
   原因：缺 import androidx.compose.animation.core.animateFloat
   修复：补 import。

2. ChatScreen.kt:402 Overload resolution ambiguity between candidates (Float.dp / Double.dp / Int.dp)
   原因：dy 是 Float，dy.dp 在 Float/Double 之间歧义
   修复：dy.toInt().dp 显式走 Int.dp 扩展。

3. ModelsScreen.kt:114 ExtendedFloatingActionButton 参数不匹配
   原因：Material3 中 ExtendedFloatingActionButton 文本/图标变体的参数顺序
        是 (text, icon, onClick, ...) 不是 (onClick, icon, text, ...)
   修复：调换参数顺序，让 onClick 在 text/icon 之后。

+++ app/src/main/kotlin/com/pocketllm/ui/chat/ChatScreen.kt
@@ -5,6 +5,7 @@ import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.InfiniteRepeatableSpec
 import androidx.compose.animation.core.RepeatMode
+import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.keyframes
 import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.animation.core.tween
@@ -399,7 +400,7 @@ private fun BouncingDot(delay: Int) {
     Box(
         Modifier
             .size(8.dp)
-            .offset(y = dy.dp)
+            .offset(y = dy.toInt().dp)
             .clip(CircleShape)
             .background(MaterialTheme.colorScheme.primary)
     )
